+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [[Unreleased]]
 
+### Fixed
+- Resolve `txnNotFound` error with `send_reliable_submission` when submitting a malformed transaction
+
 ## [1.5.0] - 2022-04-25
 ### Added
 - Support setting flags with booleans. For each transaction type supporting flags there is a `FlagInterface` to set the flags with booleans.