+++ CHANGELOG.md
@@ -6,13 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [[Unreleased]]
+Added:
+- Support for dynamic fee calculation
 
 ## [1.5.0] - 2022-04-25
 ### Added
 - Support setting flags with booleans. For each transaction type supporting flags there is a `FlagInterface` to set the flags with booleans.
 - `federator_info` RPC support
 - Helper method for creating a cross-chain payment to/from a sidechain
-- Support for dynamic fee calculation
 - Helper method for parsing an NFTokenID
 
 ### Fixed