+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 - Support for dynamic fee calculation
 
+### Fixed
+- Resolve `txnNotFound` error with `send_reliable_submission` when waiting for a submitted malformed transaction
+
 ## [1.5.0] - 2022-04-25
 ### Added
 - Support setting flags with booleans. For each transaction type supporting flags there is a `FlagInterface` to set the flags with booleans.