+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Resolve `txnNotFound` error with `send_reliable_submission` when waiting for a submitted malformed transaction
 - Small typing mistake in GenericRequest
-- Fix GenericRequest initialization bug
+- Fix bug in GenericRequest.to_dict()
 
 ## [1.5.0] - 2022-04-25
 ### Added