+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for dynamic fee calculation
 - Add function to parse account balances from a transaction's metadata
 - Better error handling for invalid client URL
+- Exported SubscribeBook
 
 ### Fixed
 - Resolve `txnNotFound` error with `send_reliable_submission` when waiting for a submitted malformed transaction