+++ CHANGELOG.md
@@ -6,10 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [[Unreleased]]
+
+## [1.6.0] - 2022-06-02
 ### Added:
 - Support for dynamic fee calculation
 - Add function to parse account balances from a transaction's metadata
 - Better error handling for invalid client URL
+- Exported SubscribeBook
 
 ### Fixed
 - Resolve `txnNotFound` error with `send_reliable_submission` when waiting for a submitted malformed transaction